+++ backend/tests/app/routers/test_threads.py
@@ -15,9 +15,7 @@ from tests.conftest import mock_keycloak_user_identification
 from tests.mock_client import MockOpenAIClient, create_mock_response
 
 
-def test_create_thread(
-    patch_required_env, httpx_mock, app_client, db_connection, test_user_info
-):
+def test_create_thread(httpx_mock, app_client, db_connection, test_user_info):
     mock_keycloak_user_identification(httpx_mock, test_user_info)
     test_settings = Settings(
         db={"prefix": db_connection}, keycloak={"issuer": "https://great_issuer.com"}
@@ -38,9 +36,7 @@ def test_create_thread(
 
 
 @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
-def test_generate_thread_title(
-    patch_required_env, httpx_mock, app_client, db_connection, test_user_info
-):
+def test_generate_thread_title(httpx_mock, app_client, db_connection, test_user_info):
     mock_keycloak_user_identification(httpx_mock, test_user_info)
     test_settings = Settings(
         db={"prefix": db_connection},
@@ -93,9 +89,7 @@ def test_generate_thread_title(
 
 
 @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
-def test_get_threads(
-    patch_required_env, httpx_mock, app_client, db_connection, test_user_info
-):
+def test_get_threads(httpx_mock, app_client, db_connection, test_user_info):
     mock_keycloak_user_identification(httpx_mock, test_user_info)
     test_settings = Settings(
         db={"prefix": db_connection}, keycloak={"issuer": "https://great_issuer.com"}
@@ -154,9 +148,7 @@ def test_get_threads(
 
 
 @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
-def test_get_threads_paginated(
-    patch_required_env, httpx_mock, app_client, db_connection, test_user_info
-):
+def test_get_threads_paginated(httpx_mock, app_client, db_connection, test_user_info):
     mock_keycloak_user_identification(httpx_mock, test_user_info)
     test_settings = Settings(
         db={"prefix": db_connection}, keycloak={"issuer": "https://great_issuer.com"}
@@ -196,9 +188,7 @@ def test_get_threads_paginated(
 
 
 @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
-def test_get_threads_query_param(
-    patch_required_env, httpx_mock, app_client, db_connection, test_user_info
-):
+def test_get_threads_query_param(httpx_mock, app_client, db_connection, test_user_info):
     mock_keycloak_user_identification(httpx_mock, test_user_info)
     test_settings = Settings(
         db={"prefix": db_connection}, keycloak={"issuer": "https://great_issuer.com"}
@@ -245,9 +235,59 @@ def test_get_threads_query_param(
 
 
 @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
-def test_update_thread_title(
-    patch_required_env, httpx_mock, app_client, db_connection, test_user_info
+def test_get_threads_exclude_empty(
+    httpx_mock,
+    app_client,
+    db_connection,
+    test_user_info,
+    populate_db,
 ):
+    mock_keycloak_user_identification(httpx_mock, test_user_info)
+    test_settings = Settings(
+        db={"prefix": db_connection}, keycloak={"issuer": "https://great_issuer.com"}
+    )
+    app.dependency_overrides[get_settings] = lambda: test_settings
+
+    _, vlab, proj = test_user_info
+    with app_client as app_client:
+        threads = app_client.get("/threads").json()
+        assert not threads["results"]
+        # Create some threads
+        create_output_1 = app_client.post(
+            "/threads", json={"virtual_lab_id": str(vlab), "project_id": str(proj)}
+        ).json()
+        create_output_2 = app_client.post(
+            "/threads", json={"virtual_lab_id": str(vlab), "project_id": str(proj)}
+        ).json()
+        create_output_3 = app_client.post(
+            "/threads", json={"virtual_lab_id": str(vlab), "project_id": str(proj)}
+        ).json()
+
+        thread_ids = [
+            create_output_1["thread_id"],
+            create_output_2["thread_id"],
+            create_output_3["thread_id"],
+        ]
+
+        # Keep empty threads
+        threads = app_client.get(
+            "/threads", params={"virtual_lab_id": vlab, "project_id": proj}
+        ).json()
+        db_items, _ = populate_db
+        assert [thread["thread_id"] for thread in threads["results"]] == list(
+            reversed([str(db_items["thread"].thread_id)] + thread_ids)
+        )
+
+        # Remove empty threads
+        threads = app_client.get(
+            "/threads",
+            params={"exclude_empty": True, "virtual_lab_id": vlab, "project_id": proj},
+        ).json()
+        assert threads["results"][0]["thread_id"] == str(db_items["thread"].thread_id)
+
+
+@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
+def test_update_thread_title(httpx_mock, app_client, db_connection, test_user_info):
     mock_keycloak_user_identification(httpx_mock, test_user_info)
     test_settings = Settings(
         db={"prefix": db_connection}, keycloak={"issuer": "https://great_issuer.com"}
@@ -282,7 +322,6 @@ def test_update_thread_title(
 
 @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 def test_delete_thread(
-    patch_required_env,
     httpx_mock,
     app_client,
     db_connection,
@@ -336,7 +375,6 @@ def test_delete_thread(
 @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 @pytest.mark.asyncio
 async def test_get_thread_messages(
-    patch_required_env,
     httpx_mock,
     app_client,
     db_connection,
@@ -386,7 +424,6 @@ async def test_get_thread_messages(
 @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 @pytest.mark.asyncio
 async def test_get_thread_messages_sort_and_filter(
-    patch_required_env,
     httpx_mock,
     app_client,
     db_connection,
@@ -455,7 +492,6 @@ async def test_get_thread_messages_sort_and_filter(
 @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 @pytest.mark.asyncio
 async def test_get_thread_messages_paginated(
-    patch_required_env,
     httpx_mock,
     app_client,
     db_connection,
@@ -514,7 +550,7 @@ async def test_get_thread_messages_paginated(
 @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 @pytest.mark.asyncio
 async def test_get_thread_messages_empty_paginated(
-    patch_required_env, httpx_mock, app_client, db_connection, test_user_info
+    httpx_mock, app_client, db_connection, test_user_info
 ):
     mock_keycloak_user_identification(httpx_mock, test_user_info)
     test_settings = Settings(
@@ -544,7 +580,6 @@ async def test_get_thread_messages_empty_paginated(
 @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
 @pytest.mark.asyncio
 async def test_get_thread_messages_vercel_format(
-    patch_required_env,
     httpx_mock,
     app_client,
     db_connection,